v2/item: Introduce `Block::size_hint`
This will be used by consumers of the crate to calculate approximate
length of the block to use.
Also move `MAX_UDP_PACKET_SIZE` from 0.1 and add a few tests.
Extracted block-related functionality into a separate submodule.
